chore(editor): editor refactoring
Fixes #162
