+++ app/plugins/modules/apache-composer/src/plugin.ts
@@ -1,5 +1,22 @@
+/*
+ * Copyright 2018 IBM Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import * as Debug from 'debug'
 const debug = Debug('plugin/apache-composer/init')
+
 import sessionList from './lib/controller/cmd/session-list'
 import sessionGet from './lib/controller/cmd/session-get'
 import invoke from './lib/controller/cmd/app-invoke'
@@ -8,6 +25,7 @@ import appCreate from './lib/controller/cmd/app-create'
 import appGet from './lib/controller/cmd/app-get'
 import appList from './lib/controller/cmd/app-list'
 import appConfig from './lib/controller/cmd/app-config'
+import editorBits from './lib/controller/cmd/editor-extensions'
 import * as usage from './usage'
 
 export default async (commandTree, prequire) => {
@@ -30,6 +48,7 @@ export default async (commandTree, prequire) => {
   await appList(commandTree, prequire)
   await appGet(commandTree, prequire)
   await appConfig(commandTree, prequire)
+  await editorBits(commandTree, prequire)
 
   debug('init done')
 }
