+++ tests/data/core/edit-file.txt
@@ -1,0 +1,1 @@
+hello world
